move mods2 to mods
