+++ src/core/analyze/mods.ts
